Implemented long mode setup/switch code, the bootloader can now start the 64-bit kernel!
The setup procedure is fairly simple: create a 64-bit GDT and 64-bit page
tables that include all kernel mappings from the 32-bit address space, but at
the correct 64-bit address, then go through kernel_args and changes all virtual
addresses to 64-bit addresses, and finally switch to long mode and jump to the
kernel.

+++ headers/private/kernel/arch/x86/arch_kernel_args.h
@@ -26,15 +26,15 @@ typedef struct {
 	uint32	system_time_cv_factor;
 	uint64	cpu_clock_speed;
 	uint32	phys_pgdir;
-	uint32	vir_pgdir;
+	uint64	vir_pgdir;
 	uint32	num_pgtables;
 	uint32	pgtables[MAX_BOOT_PTABLES];
-	uint32	virtual_end;
+	uint64	virtual_end;
 	uint32	phys_idt;
-	uint32	vir_idt;
+	uint64	vir_idt;
 	uint32	phys_gdt;
-	uint32	vir_gdt;
-	uint32	page_hole;
+	uint64	vir_gdt;
+	uint64	page_hole;
 	// smp stuff
 	uint32	apic_time_cv_factor; // apic ticks per second
 	uint32	apic_phys;
